emr-server/requirements.txt
asgiref==3.7.2
boto3==1.34.2
botocore==1.34.2
certifi==2023.11.17
cffi==1.16.0
contourpy==1.1.1
cryptography==41.0.7
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
h5py==3.9.0
importlib-resources==6.0.1
jmespath==1.0.1
numpy==1.26.2
packaging==23.2
Pillow==10.0.1
protobuf==4.25.1
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.9.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
whitenoise==6.6.0
zipp==3.17.0
